Life_In_Words/requirement.txt
Package                       Version
----------------------------- --------
Django                        5.2.3
django-debug-toolbar          6.0.0
djangorestframework           3.16.0
djangorestframework_simplejwt 5.5.1
djoser                        2.3.3
Faker                         37.4.2
mysqlclient                   2.2.7
pillow                        11.3.0
pip                           25.1.1
